refactor: merge `monkey.ts` into `gm-variables.ts`
