+++ src/components/Galleries/GalleriesEnhancer.vue
@@ -7,9 +7,8 @@ import { VueFinalModal } from 'vue-final-modal'
 import { setWheelStep } from '@/utils/wheel-step'
 import { usePopups } from '@/composables/usePopups'
 import { getDoc, getElement, getElements } from '@/utils/commons'
-import { scrollByRowSwitch, infiniteScrollSwitch, archiveButtonSwitch, quickArchiveDownloadMethod } from '@/utils/gm-variables'
+import { scrollByRowSwitch, infiniteScrollSwitch, archiveButtonSwitch, quickArchiveDownloadMethod, ArchiveDownloadMethod } from '@/utils/gm-variables'
 import { getArchiveLink } from '@/utils/e-hentai-api'
-import { ArchiveDownloadMethod } from '@/constants/monkey'
 import { useArchive } from '@/composables/useArchive'
 import { highlightDownloadedGalleries, watchDownloadedGalleries } from '@/utils/highlight-galleries'
