+++ src/components/SettingsPanel/CrossButton.vue
