--- src/constants/monkey.ts
@@ -1,40 +1,0 @@
-export const enum GMKey {
-  InfiniteScroll = 'InfiniteScroll',
-  ScrollByRow = 'ScrollByRow',
-  ArchiveButton = 'archiveButton',
-  Highlight = 'Highlight',
-
-  BetterPopup = 'BetterPopup',
-  QuickArchiveDownloadMethod = 'QuickDownloadMethod',
-  QuickTorrentDownload = 'QuickTorrentDownload',
-  LoadAllGalleryImages = 'LoadAllGalleryImages',
-
-  MultipageViewerEnhancer = 'MultipageViewerEnhancer',
-  AutoRedirect = 'AutoRedirect',
-  PreventImageRemoval = 'PreventImageRemoval',
-  ChangePageByWheelAnyWhereSwitch = 'ChangePageByWheelAnyWhere',
-
-  Magnifier = 'Magnifier',
-  MagnifierActivationButton = 'MagnifierActivationButton',
-  MagnifierToggleMode = 'MagnifierToggleMode',
-  MagnifierDefaultScale = 'MagnifierDefaultScale',
-  MagnifierScaleStep = 'MagnifierScaleStep',
-  MagnifierLongPressThreshold = 'MagnifierLongPressThreshold',
-  MagnifierSensitivityX = 'MagnifierSensitivityX',
-  MagnifierSensitivityY = 'MagnifierSensitivityY',
-
-  ShowJapaneseTitle = 'ShowJapaneseTitle',
-}
-
-export const enum ArchiveDownloadMethod {
-  Manual = 'Manual',
-  HaH_Original = 'download Original Resolution with H@H',
-  HaH_2400 = 'download 2400x Resolution with H@H',
-  Direct_Origin = 'download Original Resolution directly',
-  Direct_Resample = 'download Resample Resolution directly',
-}
-
-export const enum MouseButton {
-  Left = 'left',
-  Right = 'right',
-}
